+++ enlighten/Controller.py
@@ -1804,7 +1804,7 @@ class Controller:
         # lesser value and reserved 0xffff for "frame markers". Raw should match 
         # processed at this point, but "raw" is preferred to clarify saturation 
         # is irrespective of dark correction or any post-processing.
-        if pr.raw.max() >= 0xfffe: 
+        if pr.raw.max() >= 0xfffe and not spec.settings.is_ids():
             # todo: self.status_indicators.detector_warning("detector saturated")
             self.marquee.error("detector saturated")