+++ plugins/Network/IDSPeak.py
@@ -19,8 +19,7 @@ class IDSPeak(EnlightenPluginBase):
         choices = IDSCamera.SUPPORTED_CONVERSIONS
 
         # self.field(name="Connect", direction="input", datatype="button", callback=self.connect_callback, tooltip="Attempt to connect to an IDS camera via the IDSPeak SDK")
-        self.field(name="Vertical Binning", direction="input", datatype="combobox", choices=choices, callback=self.format_callback, tooltip="Image format used for vertical binning")
-        self.field(name="Area Scan", direction="input", datatype="combobox", choices=choices, callback=self.format_callback, tooltip="Image format used for area scan")
+        self.field(name="Format", direction="input", datatype="combobox", choices=choices, callback=self.format_callback, tooltip="Image format used for vertical binning")
 
         device_id = DeviceID(label="IDSPeak")
         log.debug(f"adding to other_device_ids: {device_id}")
@@ -35,8 +34,5 @@ class IDSPeak(EnlightenPluginBase):
             self.ctl.marquee.error("IDSPeak plugin can only be used with IDSPeak spectrometers")
             return
 
-        vert_name = self.get_widget_from_name("Vertical Binning").currentText()
-        area_name = self.get_widget_from_name("Area Scan").currentText()
-
-        spec.change_device_setting("vertical_binning_format_name", vert_name)
-        spec.change_device_setting("area_scan_format_name", area_name)
+        format_name = self.get_widget_from_name("Format").currentText()
+        spec.change_device_setting("vertical_binning_format_name", format_name)