+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Version History
 
-- 2025-??-?? 4.1.25
+- 2025-04-04 4.1.25
     - IDS
         - added IDSPeak plugin
         - update AreaScan to support .png files
     - Area Scan
         - deprecated 'Fast' checkbox
         - added 'Normalize' checkbox
+    - plugins
+        - fixed Prod.EmissionLines unit
 - 2025-02-28 4.1.24
     - plugins
         - automatically add "Copy to Clipboard" button for plugins generating a dataframe